now you can update category

+++ admin/functions/categories/c_manager.php
@@ -12,18 +12,32 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
             // IS [action] variable is set ?
             if (isset($_GET["action"])) {
                 if (!empty($_POST['category'])) {
-                    try {
-                        $stm = $dbc->prepare('INSERT into category (name) values(:category)'); // add category
-                        $stm->bindParam(":category", $_POST["category"]);
-                        $stm->execute();
-                        // $data = $stm ->fetchAll(PDO::FETCH_ASSOC);
-                        if ($stm->rowcount() > 0) {
-                            header("Location:" . "../../pages/categories.php?statu=1");
+                    if($_GET['action'] == 3){ // 3 =   add
+                        try {
+                            $stm = $dbc->prepare('INSERT into category (name) values(:category)'); // add category
+                            $stm->bindParam(":category", $_POST["category"]);
+                            $stm->execute();
+                            // $data = $stm ->fetchAll(PDO::FETCH_ASSOC);
+                            if ($stm->rowcount() > 0) {
+                                header("Location:" . "../../pages/categories.php?statu=1");
+                            }
+                        } catch (Exception $e) {
+                            // echo $e;
+                            header("Location:" . "../../pages/categories.php?statu=0");
                         }
-                    } catch (Exception $e) {
-                        // echo $e;
-                        header("Location:" . "../../pages/categories.php?statu=0");
+                    }elseif($_GET["action"] == 1){ // 1 = update
+                        try {
+                            $stm = $dbc->prepare('update  category set name = :name where id = :id'); // add category
+                            $stm->bindParam(":id", $_GET["id"]);
+                            $stm->bindParam(":name", $_POST["category"]);
+                            $stm->execute();
+                            header("Location:" . "../../pages/categories.php?statu=1");
+                        } catch (Exception $e) {
+                             //echo $e;
+                            header("Location:" . "../../pages/categories.php?statu=0");
+                        } 
                     }
+
                 }
             }
         }
@@ -36,15 +50,17 @@ if ($_SERVER['REQUEST_METHOD'] == 'GET') {
         if ($_SESSION["privilege"] == 1) {
             // IS [action] variable is set ?
             if (isset($_GET["action"]) && isset($_GET["id"])) {
-                try {
-                    $stm = $dbc->prepare('delete from category where id = :id'); // add category
-                    $stm->bindParam(":id", $_GET["id"]);
-                    $stm->execute();
-                    header("Location:" . "../../pages/categories.php?statu=1");
-                } catch (Exception $e) {
-                    // echo $e;
-                    header("Location:" . "../../pages/categories.php?statu=0");
-                } 
+                    if($_GET["action"] == 2){
+                        try {
+                        $stm = $dbc->prepare('delete from category where id = :id'); // add category
+                        $stm->bindParam(":id", $_GET["id"]);
+                        $stm->execute();
+                        header("Location:" . "../../pages/categories.php?statu=1");
+                    } catch (Exception $e) {
+                        //echo $e;
+                        header("Location:" . "../../pages/categories.php?statu=0");
+                    } 
+                }
             }
         }
     }
